requirements.txt
# WordPress Traffic Generator Dependencies
playwright>=1.40.0
langchain-ollama>=0.1.0
langchain-core>=0.1.0
matplotlib>=3.8.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
